Refactor TowerComponent and VelocityComponent classes, and update AudioPack class.

+++ core/src/com/softwarearchitecture/ecs/components/TowerComponent.java
@@ -6,17 +6,16 @@ public class TowerComponent implements Serializable {
     // TODO: Change all fields to public and remove getters and setters
     private int damage;
     private float range;
-    private int attackCooldown;
+    private float attackCooldown;
     private float timeSinceLastAttack;
     public boolean playSound = false;
 
-    public TowerComponent(int damage, float range, int attackCooldown) {
+    public TowerComponent(int damage, float range, float attackCooldown) {
         if (damage >= 1 && range >= 1 && attackCooldown >= 1) {
             this.damage = damage;
             this.range = range;
             this.attackCooldown = attackCooldown;
-        }
-        else {
+        } else {
             this.damage = 1;
             this.range = 1;
             this.attackCooldown = 1;
@@ -34,8 +33,7 @@ public class TowerComponent implements Serializable {
     public void setDamage(int damage) {
         if (damage >= 1) {
             this.damage = damage;
-        }
-        else {
+        } else {
             this.damage = 1;
         }
     }
@@ -43,8 +41,7 @@ public class TowerComponent implements Serializable {
     public void setRange(int range) {
         if (range >= 1) {
             this.range = range;
-        }
-        else {
+        } else {
             this.range = 1;
         }
     }
